ex 4.21

+++ part4/blog/controllers/blogs.js
@@ -36,8 +36,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    user.blogs = user.blogs.filter(b => b.toString() !== blog.id )
+    await user.save()
+    return response.status(204).end()
+  } 
+  
+  return response.status(401).json({ error: 'token missing or invalid'})
 })
 
 blogsRouter.put('/:id', async (request, response) => {
